Pearson kernels are here!
At least a beta version of it is. The iprior function now recognises
categorical variables in the dataframe (as factors) and applied the
Pearson kernel appropriately. All other variables are still applied the
Canonical kernel.

This commit fixes issues #2, and then some. More (hopefully useful)
information is added into the summary.