Further enhancements to Pearson kernel
Previous version was wrong for several reason:
1. The Pearson kernel function `fn.H1` did not correctly concatenate
factors. Must use `unlist(list(factor1, factor2))` instead of `c`. As a
result, the input of `fn.H1(x,y)` was incorrect, but `fn.H1(x)` was OK.
2. The `iprior.predict` function did not use the correct kernel
function, so was not able to produce the correct kernel for prediction.
This is fixed.

Moreover, the Pearson kernel function was re-written and now is
massively faster than the previous version.