Fixed #6 - Kernel functions
There was an error whereby prediction of existing data points were not
equivalent to the fitted values. The error was in the Pearson kernel
(`p_i` musn’t change when new data point comes in), and in the
Canonical kernel (wrong mean used, so the kernel was centred wrongly).

Also made some tweaks to the random starting values for `psi` (made
them tighter around 0, as it is most likely that psi will be small -
though this is an unproven conjecture).