Massive changes under the hood
Here, I have done away with three separate R files and combined all 3
types of EM into a single R file. The reason is that previously, I had
to change 3 separate files because the code was repeated in 3 separate
files. Now there should be less work to do when I need to implement
future updates such as the predicted log-likelihood or changes to the
EM reporting interface.

Also, some performance enhancement to the EM (issue #20). So far, no.1
and no.2 has been addressed.