This is a massive update.
1. The main EM engine has now been entirely re-written for speed and
efficiency. Some details are described in the conversation of issue
#20.

2. New plotting feature. This gives a fitted curve (only if it fits in
a 2-D y vs x plot), residuals and Q-Q plot of residuals.

3. The calculation of standard errors now make use of the final EM
values using closed form formulas. Significantly faster than the old
way of re-calculating the Hessian using a function from an outside
package.

4. Added some preliminary test functions.

5. There is also a `beta()` function which can recover the betas if a
Canonical kernel is used. Still work to be done on this function.